bloglist (exs 4.18-4.22)

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
+const middleware = require('../utils/middleware')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
@@ -11,14 +13,13 @@ blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
     .populate('user', { username: 1, name: 1 })
-    
+
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-
-  const user = await User.findById(body.userId)
+  const user = await User.findById(request.userId)
   if (!user) {
     return response.status(400).json({ error: 'userId missing or not valid' })
   }
@@ -48,10 +49,22 @@ blogsRouter.put('/:id', async (request, response) => {
   blog.likes = likes
   const result = await blog.save()
   response.status(204).json(result)
-
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(400).json({ error: 'blog not found' })
+  }
+  const user = await User.findById(request.userId)
+  if (!user) {
+    return response.status(400).json({ error: 'user not found' })
+  }
+
+  if (blog.user.toString() !== user.id) {
+    return response.status(400).json({ error: 'user is not owner of blog' })
+  }
+  
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
